requirements.txt
fastapi
uvicorn
requests
folium
geopy
aiofiles
pandas
google-generativeai
